requirements.txt
aiohttp
faker
web3
aiohttp_cors
schedule
requests
PyGithub
retry